+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   Debugger address (`debugger_address`), java path (`java_path`) and flag for trace log (`produce_trace_log`) are *no 
   longer* fields in project view files! They can be set *only* using installer flags (check [README](install/README.md)).
 - Project view `bazel_path` has been renamed to `bazel_binary` (now it's compatible with https://ij.bazel.build/docs/project-views.html#bazel_binary).
+- Project view `targets` has its default value changed from all targets (`//...`) to no targets. | [#BAZEL-552](https://youtrack.jetbrains.com/issue/BAZEL-552).
 
 ### Features 🎉
 - Enhance support for Kotlin by providing Kotlin target's specific info.